fix merge conflict

+++ src/main/java/com/idle/kb_i_dle_backend/domain/finance/dto/FinancialSumDTO.java
@@ -1,14 +1,15 @@
 package com.idle.kb_i_dle_backend.domain.finance.dto;
 
+import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
 public class FinancialSumDTO {
-    private long amount;  // 금융 자산 합계
+    private BigDecimal amount;  // 금융 자산 합계
 
-    public FinancialSumDTO(long amount) {
+    public FinancialSumDTO(BigDecimal amount) {
         this.amount = amount;
     }
 }
