+++ src/main/java/com/idle/kb_i_dle_backend/domain/goal/dto/OutcomeGoalDTO.java
@@ -1,6 +1,6 @@
 package com.idle.kb_i_dle_backend.domain.goal.dto;
 
-import lombok.AllArgsConstructor;
+import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -10,14 +10,14 @@ import java.util.Date;
 @Setter
 public class OutcomeGoalDTO {
     private int index;
-    private long gather;
+    private BigDecimal gather;
     private String title;
-    private long amount;
+    private BigDecimal amount;
     private Date date;
     private int priority;
 
 
-    public OutcomeGoalDTO(int index, long gather, String title, long amount, Date date, int priority) {
+    public OutcomeGoalDTO(int index, BigDecimal gather, String title, BigDecimal amount, Date date, int priority) {
         this.index = index;
         this.gather = gather;
         this.title = title;
@@ -25,4 +25,6 @@ public class OutcomeGoalDTO {
         this.date = date;
         this.priority = priority;
     }
+
+
 }
